javaSpringMVC/src/test/java/com/javamvcweb/javaSpringMVC/controller/UserControllerTests.java
package com.javamvcweb.javaSpringMVC.controller;

import com.javamvcweb.controller.UserController;
import com.javamvcweb.service.FormService;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.ui.Model;

import static org.mockito.Mockito.*;
import static org.junit.jupiter.api.Assertions.*;

class UserControllerTests {

    @Mock
    private FormService formService;

    @InjectMocks
    private UserController userController;

    @Test
    void testShowForm() {
        Model model = mock(Model.class);

        // Mock the behavior of the formService if needed
        // when(formService.someMethod()).thenReturn(someValue);

        String result = userController.showForm(model);

        // Add assertions based on your business logic
        assertEquals("user-form", result); // Check that the correct view name is returned

        // Verify that the formService method was called if applicable
        // verify(formService, times(1)).someMethod();

        // Add more assertions as needed
    }

    // Add more tests as needed
}
